requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
platformdirs==4.3.7
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.29.3
